requirements.txt
Flask==3.0.2
GitPython==3.1.41 